refactor: remove explicit timestamp verification in job timeout checker test
The test cases in the JobTimeoutCheckerTest class verify the invocation
of the task scheduling method. Inside the JobTimeoutChecker class, the
runDelayed method was replaced with the runAt method, which requires an
exact timestamp. This replacement makes it more difficult to verify the
exact value provided by the ActorClock which is used for the task
scheduling. However, the test cases are structured in such a way that
the timestamp verification doesn't impact the correctness of the test
so we can ignore it.